requirements.txt
langchain>=0.1.17
qdrant-client
openai
pypdf
python-dotenv
flask
-e .